Abstract prefixed command parsers further
Previously when CommandWithPrefixParser was generalised out of
AddCommandParser, it catered to the needs of commands that use prefixes,
out of which some are compulsory prefixes and has preamble. Some
commands like edit do not have compulsory prefixes or preamble, thus it
is necessary to further simplify it for such use cases.
